tools/set-up-repositories/requirements.txt
gitpython~=3.1.44
PyGithub~=2.6.1

requests~=2.32.4
python-dotenv~=1.1.0
PyJWT~=2.10.1
